+++ package.json
@@ -23,6 +23,10 @@
     "README.md"
   ],
   "author": "David Wells",
+  "repository": "https://github.com/netlify-labs/netlify-plugin-sitemap",
+  "bugs": {
+    "url": "https://github.com/netlify-labs/netlify-plugin-sitemap/issues"
+  },
   "license": "MIT",
   "dependencies": {
     "globby": "^10.0.1",
